+++ bin/build-cultures.php
@@ -32,7 +32,11 @@ $url    = 'https://msdn.microsoft.com/en-us/library/ee825488(v=cs.20).aspx';
 $buffer = file_get_contents($url);
 $values = [];
 
-if (preg_match_all('/<td data-th="Language Culture Name">(.*)<\/td>/U', $buffer, $matches) && isset($matches[1])) {
+if (preg_match_all(
+        '/<td data-th="Language Culture Name">(.*)<\/td>/U',
+        $buffer,
+        $matches
+    ) && isset($matches[1])) {
     $values = array_map('trim', $matches[1]);
 }