+++ src/DeleteTrait.php
@@ -1,0 +1,54 @@
+<?php
+
+/**
+ * ReportingCloud PHP Wrapper
+ *
+ * PHP wrapper for ReportingCloud Web API. Authored and supported by Text Control GmbH.
+ *
+ * @link      http://www.reporting.cloud to learn more about ReportingCloud
+ * @link      https://github.com/TextControl/txtextcontrol-reportingcloud-php for the canonical source repository
+ * @license   https://raw.githubusercontent.com/TextControl/txtextcontrol-reportingcloud-php/master/LICENSE.md
+ * @copyright © 2017 Text Control GmbH
+ */
+
+namespace TxTextControl\ReportingCloud;
+
+use GuzzleHttp\Psr7\Response;
+use GuzzleHttp\RequestOptions;
+use TxTextControl\ReportingCloud\Exception\InvalidArgumentException;
+use TxTextControl\ReportingCloud\Validator\StaticValidator;
+
+trait DeleteTrait
+{
+    /**
+     * Delete a template in template storage
+     *
+     * @param string $templateName Template name
+     *
+     * @throws InvalidArgumentException
+     *
+     * @return bool
+     */
+    public function deleteTemplate($templateName)
+    {
+        $ret = false;
+
+        StaticValidator::execute($templateName, 'TemplateName');
+
+        $query = [
+            'templateName' => $templateName,
+        ];
+
+        $options = [
+            RequestOptions::QUERY => $query,
+        ];
+
+        $response = $this->request('DELETE', $this->uri('/templates/delete'), $options);
+
+        if ($response instanceof Response && 204 === $response->getStatusCode()) {
+            $ret = true;
+        }
+
+        return $ret;
+    }
+}