+++ src/PutTrait.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * ReportingCloud PHP Wrapper
+ *
+ * PHP wrapper for ReportingCloud Web API. Authored and supported by Text Control GmbH.
+ *
+ * @link      http://www.reporting.cloud to learn more about ReportingCloud
+ * @link      https://github.com/TextControl/txtextcontrol-reportingcloud-php for the canonical source repository
+ * @license   https://raw.githubusercontent.com/TextControl/txtextcontrol-reportingcloud-php/master/LICENSE.md
+ * @copyright © 2017 Text Control GmbH
+ */
+
+namespace TxTextControl\ReportingCloud;
+
+trait PutTrait
+{
+    /* coming soon */
+}