+++ test/ReportingCloudTest.php
@@ -2,6 +2,8 @@
 
 namespace TxTextControlTest\ReportingCloud;
 
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\ClientException;
 use PHPUnit_Framework_Constraint_IsType as PHPUnit_IsType;
 use PHPUnit_Framework_TestCase;
 use TxTextControl\ReportingCloud\Console\Helper;
@@ -25,7 +27,84 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $this->reportingCloud->setPassword(Helper::password());
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
+    public function testConstructorOptions()
+    {
+        $options = [
+            'username' => 'phpunit-username',
+            'password' => 'phpunit-password',
+            'base_uri' => 'https://api.example.com',
+            'timeout'  => 100,
+            'version'  => 'v1',
+            'debug'    => true,
+            'test'     => true,
+        ];
+
+        $reportingCloud = new ReportingCloud($options);
+
+        $this->assertSame('phpunit-username', $reportingCloud->getUsername());
+        $this->assertSame('phpunit-password', $reportingCloud->getPassword());
+        $this->assertSame('https://api.example.com', $reportingCloud->getBaseUri());
+        $this->assertSame(100, $reportingCloud->getTimeout());
+        $this->assertSame('v1', $reportingCloud->getVersion());
+
+        $this->assertTrue($reportingCloud->getDebug());
+        $this->assertTrue($reportingCloud->getTest());
+
+        unset($reportingCloud);
+    }
+
+    public function testSetGetProperties()
+    {
+        $this->reportingCloud->setUsername('phpunit-username');
+        $this->reportingCloud->setPassword('phpunit-password');
+        $this->reportingCloud->setBaseUri('https://api.example.com');
+        $this->reportingCloud->setTimeout(100);
+        $this->reportingCloud->setVersion('v1');
+        $this->reportingCloud->setDebug(true);
+        $this->reportingCloud->setTest(true);
+
+        $this->assertSame('phpunit-username', $this->reportingCloud->getUsername());
+        $this->assertSame('phpunit-password', $this->reportingCloud->getPassword());
+        $this->assertSame('https://api.example.com', $this->reportingCloud->getBaseUri());
+        $this->assertSame(100, $this->reportingCloud->getTimeout());
+        $this->assertSame('v1', $this->reportingCloud->getVersion());
+
+        $this->assertTrue($this->reportingCloud->getDebug());
+        $this->assertTrue($this->reportingCloud->getTest());
+    }
+
+    public function testGetClientInstanceOf()
+    {
+        $this->assertInstanceOf('GuzzleHttp\Client', $this->reportingCloud->getClient());
+    }
+
+    public function testDefaultProperties()
+    {
+        $this->assertNull($this->reportingCloud->getUsername());
+        $this->assertNull($this->reportingCloud->getPassword());
+
+        $this->assertSame('https://api.reporting.cloud', $this->reportingCloud->getBaseUri());
+        $this->assertSame(120, $this->reportingCloud->getTimeout());
+        $this->assertSame('v1', $this->reportingCloud->getVersion());
+
+        $this->assertFalse($this->reportingCloud->getDebug());
+    }
+
+    public function testResponseStatusCodeOnNonSsl()
+    {
+        $baseUriHost = parse_url($this->reportingCloud->getBaseUri(), PHP_URL_HOST);
+
+        $this->assertNotEmpty($baseUriHost);
+
+        $uri = sprintf('http://%s', $baseUriHost);
+
+        try {
+            $client = new Client();
+            $client->request('GET', $uri);
+        } catch (ClientException $e) {
+            $this->assertSame(404, $e->getResponse()->getStatusCode());
+        }
+    }
 
     public function testProofingCheck()
     {
@@ -48,11 +127,9 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('en_US.dic', $response['language']);
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     public function testGetAvailableDictionaries()
     {
-        $filename = realpath(__DIR__ . '/../resource/dictionaries.php');
+        $filename = realpath(__DIR__ . '/../data/dictionaries.php');
 
         $actual   = $this->reportingCloud->getAvailableDictionaries();
         $expected = include $filename;
@@ -63,8 +140,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $this->assertEquals($expected, $actual);
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     public function testGetProofingSuggestions()
     {
         $response = $this->reportingCloud->getProofingSuggestions('Thiss', 'en_US.dic', 5);
@@ -73,8 +148,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         $this->assertContains('This', $response);
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     public function testGetTemplateInfo()
     {
@@ -130,11 +203,9 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         unlink($tempTemplateFilename);
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     protected function getTestTemplateFilename()
     {
-        $ret = sprintf('%s/test_template.tx', realpath(__DIR__ . '/TestAsset'));
+        $ret = sprintf('%s/test_template.tx', realpath(__DIR__ . '/../data'));
 
         return $ret;
     }
@@ -197,8 +268,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $this->reportingCloud->getTemplateThumbnails('sample_invoice.tx', 100, -1, 1, 'PNG');
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     /**
      * @expectedException InvalidArgumentException
      */
@@ -206,8 +275,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
     {
         $this->reportingCloud->getTemplateThumbnails('sample_invoice.tx', 100, 1, -1, 'PNG');
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     /**
      * @expectedException InvalidArgumentException
@@ -245,8 +312,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         unlink($tempTemplateFilename);
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     /**
      * @expectedException InvalidArgumentException
@@ -287,8 +352,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         unlink($tempTemplateFilename);
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     /**
      * @expectedException InvalidArgumentException
      */
@@ -296,8 +359,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
     {
         $this->reportingCloud->getTemplatePageCount('sample_invoice.xx');
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     public function testGetTemplatePageCount()
     {
@@ -325,8 +386,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         unlink($tempTemplateFilename);
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     public function testGetTemplateList()
     {
@@ -362,8 +421,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         unlink($tempTemplateFilename);
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     public function testGetFontList()
     {
@@ -431,8 +488,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $this->reportingCloud->convertDocument('/invalid/path/document.doc', 'PDF');
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     /**
      * @expectedException InvalidArgumentException
      */
@@ -459,7 +514,7 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
     protected function getTestDocumentFilename()
     {
-        $ret = sprintf('%s/test_document.docx', realpath(__DIR__ . '/TestAsset'));
+        $ret = sprintf('%s/test_document.docx', realpath(__DIR__ . '/../data'));
 
         return $ret;
     }
@@ -617,8 +672,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $this->reportingCloud->mergeDocument($mergeData, 'PDF', null, $templateFilename, true, 1);
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     /**
      * @expectedException InvalidArgumentException
      */
@@ -717,7 +770,14 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         foreach ($returnFormats as $returnFormat) {
 
-            $response = $this->reportingCloud->mergeDocument($mergeData, $returnFormat, $tempTemplateName, null, false, $mergeSettings);
+            $response = $this->reportingCloud->mergeDocument(
+                $mergeData,
+                $returnFormat,
+                $tempTemplateName,
+                null,
+                false,
+                $mergeSettings
+            );
 
             $this->assertNotNull($response);
             $this->assertNotFalse($response);
@@ -756,7 +816,14 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         foreach ($returnFormats as $returnFormat) {
 
-            $response = $this->reportingCloud->mergeDocument($mergeData, $returnFormat, null, $testTemplateFilename, false, $mergeSettings);
+            $response = $this->reportingCloud->mergeDocument(
+                $mergeData,
+                $returnFormat,
+                null,
+                $testTemplateFilename,
+                false,
+                $mergeSettings
+            );
 
             $this->assertNotNull($response);
             $this->assertNotFalse($response);
@@ -819,8 +886,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $this->reportingCloud->findAndReplaceDocument($findAndReplaceData, 'PDF', null, '/invalid/path/template');
     }
 
-    // -----------------------------------------------------------------------------------------------------------------
-
     /**
      * @expectedException InvalidArgumentException
      */
@@ -856,7 +921,7 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
     protected function getTestTemplateFindAndReplaceFilename()
     {
-        $ret = sprintf('%s/test_find_and_replace.tx', realpath(__DIR__ . '/TestAsset'));
+        $ret = sprintf('%s/test_find_and_replace.tx', realpath(__DIR__ . '/../data'));
 
         return $ret;
     }
@@ -877,7 +942,13 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $mergeSettings['remove_empty_fields'] = 'invalid';
         $mergeSettings['remove_empty_images'] = 'invalid';
 
-        $this->reportingCloud->findAndReplaceDocument($findAndReplaceData, 'PDF', null, $templateFilename, $mergeSettings);
+        $this->reportingCloud->findAndReplaceDocument(
+            $findAndReplaceData,
+            'PDF',
+            null,
+            $templateFilename,
+            $mergeSettings
+        );
     }
 
     /**
@@ -895,7 +966,13 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
         $mergeSettings['creation_date']          = -1;  // value must be timestamp
         $mergeSettings['last_modification_date'] = 'invalid';
 
-        $this->reportingCloud->findAndReplaceDocument($findAndReplaceData, 'PDF', null, $templateFilename, $mergeSettings);
+        $this->reportingCloud->findAndReplaceDocument(
+            $findAndReplaceData,
+            'PDF',
+            null,
+            $templateFilename,
+            $mergeSettings
+        );
     }
 
     public function testFindAndReplaceDocumentWithTemplateName()
@@ -923,7 +1000,13 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         foreach ($returnFormats as $returnFormat) {
 
-            $response = $this->reportingCloud->findAndReplaceDocument($findAndReplaceData, $returnFormat, $tempTemplateName, null, $mergeSettings);
+            $response = $this->reportingCloud->findAndReplaceDocument(
+                $findAndReplaceData,
+                $returnFormat,
+                $tempTemplateName,
+                null,
+                $mergeSettings
+            );
 
             $this->assertNotNull($response);
             $this->assertNotFalse($response);
@@ -934,8 +1017,6 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         $this->assertTrue($response);
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     public function testFindAndReplaceDocumentWithTemplateFilename()
     {
@@ -950,15 +1031,19 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         foreach ($returnFormats as $returnFormat) {
 
-            $response = $this->reportingCloud->findAndReplaceDocument($findAndReplaceData, $returnFormat, null, $testTemplateFilename, $mergeSettings);
+            $response = $this->reportingCloud->findAndReplaceDocument(
+                $findAndReplaceData,
+                $returnFormat,
+                null,
+                $testTemplateFilename,
+                $mergeSettings
+            );
 
             $this->assertNotNull($response);
             $this->assertNotFalse($response);
             $this->assertGreaterThanOrEqual(1024, mb_strlen($response));
         }
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
 
     /**
      * @expectedException InvalidArgumentException
@@ -997,7 +1082,8 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
      */
     public function testDownloadTemplateInvalidTemplateName()
     {
-        $templateFilename = $this->getTestTemplateFilename();   // should be templateName and not templateFilename
+        // should be templateName and not templateFilename
+        $templateFilename = $this->getTestTemplateFilename();
 
         $this->reportingCloud->downloadTemplate($templateFilename);
     }
@@ -1007,7 +1093,8 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
      */
     public function testDeleteTemplateInvalidTemplateName()
     {
-        $templateFilename = $this->getTestTemplateFilename();   // should be templateName and not templateFilename
+        // should be templateName and not templateFilename
+        $templateFilename = $this->getTestTemplateFilename();
 
         $this->reportingCloud->deleteTemplate($templateFilename);
     }
@@ -1056,7 +1143,4 @@ class ReportingCloudTest extends PHPUnit_Framework_TestCase
 
         return $ret;
     }
-
-    // -----------------------------------------------------------------------------------------------------------------
-
 }