+++ test/TestAsset/ConcreteReportingCloud.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace TxTextControlTest\ReportingCloud\TestAsset;
-
-use TxTextControl\ReportingCloud\AbstractReportingCloud;
-
-class ConcreteReportingCloud extends AbstractReportingCloud
-{
-}